Fix on whether to serialize tda limits or load it

+++ package/database.py
@@ -131,37 +131,45 @@ class Database:
     # Build the corresponding Betti curves
     def add_betti_curves(self):
 
-        lmt = []
-        # Get the betti curves limitations
-        for pth in [self.train_out, self.valid_out]:
+        pth = './dataset/TDA_limits.pk'
+        # Limitations of persistence diagrams
+        if os.path.exists(pth):
 
-            # Iterates over the EEGs signals
-            for key in tqdm.tqdm(range(1, 5)):
+            # Load the existing thresholds
+            with open(, 'rb') as raw: dic = pickle.load(raw)
 
-                # Load the corresponding values
-                with h5py.File(pth, 'r') as dtb: 
-                    val = dtb['eeg_{}'.format(key)].value
+        else:
 
-                # Computes the persistent limits for the relative patient
-                pol = multiprocessing.Pool(processes=self.threads)
-                lmt.append(np.asarray(pol.map(persistent_limits, val)))
-                pol.close()
-                pol.join()
-                # Memory efficiency
-                del val, pol
+            lmt = []
+            # Get the betti curves limitations
+            for pth in [self.train_out, self.valid_out]:
 
-        # Extracts the main limits
-        lmt = np.vstack(tuple(lmt))
-        mnu, mxu = min(lmt[:,0]), max(lmt[:,1]) 
-        mnd, mxd = min(lmt[:,2]), max(lmt[:,3])
-        # Memory efficiency
-        del lmt
+                # Iterates over the EEGs signals
+                for key in tqdm.tqdm(range(1, 5)):
 
-        # Serialize the obtained threshold
-        with open('./tools/TDA_limits.pk', 'wb') as raw:
-            dic = {'min_up': mnu, 'max_up': mxu, 'min_dw': mnd, 'max_dw': mxd}
-            pickle.dump(dic, raw)
-            del dic
+                    # Load the corresponding values
+                    with h5py.File(pth, 'r') as dtb: 
+                        val = dtb['eeg_{}'.format(key)].value
+
+                    # Computes the persistent limits for the relative patient
+                    pol = multiprocessing.Pool(processes=self.threads)
+                    lmt.append(np.asarray(pol.map(persistent_limits, val)))
+                    pol.close()
+                    pol.join()
+                    # Memory efficiency
+                    del val, pol
+
+            # Extracts the main limits
+            lmt = np.vstack(tuple(lmt))
+            mnu, mxu = min(lmt[:,0]), max(lmt[:,1]) 
+            mnd, mxd = min(lmt[:,2]), max(lmt[:,3])
+            # Memory efficiency
+            del lmt
+
+            # Serialize the obtained threshold
+            with open(, 'wb') as raw:
+                dic = {'min_up': mnu, 'max_up': mxu, 'min_dw': mnd, 'max_dw': mxd}
+                pickle.dump(dic, raw)
 
         # Build the betti curves
         for pth in [self.train_out, self.valid_out]:
@@ -175,7 +183,7 @@ class Database:
                     
                 # Multiprocessed computation
                 pol = multiprocessing.Pool(processes=self.threads)
-                arg = {'mnu': mnu, 'mxu': mxu, 'mnd': mnd, 'mxd': mxd}
+                arg = {'mnu': dic['mnu'], 'mxu': dic['mxu'], 'mnd': dic['mnd'], 'mxd': dic['mxd']}
                 fun = partial(compute_betti_curves, **arg)
                 res = np.asarray(pol.map(fun, val))
                 pol.close()
